pyproject.toml
[project]
name = "guardian-angel"
version = "0.2.1"
description = "Policy engine for governing AI agent tool execution."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Po-Yao Huang", email = "poyaohg0705@gmail.com" },
]
keywords = ["agents", "policy", "ai", "governance", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cli = [
    "rich>=14.3.3",
    "typer>=0.24.1",
]
test = ["pytest>=7.0"]

[project.scripts]
guardian-angel = "guardian_angel.cli:main"

[project.urls]
Homepage = "https://github.com/poyao0705/guardian-angel"
Repository = "https://github.com/poyao0705/guardian-angel"
Issues = "https://github.com/poyao0705/guardian-angel/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guardian_angel"]
